requirements.txt
fastapi==0.85.0
passlib==1.7.4
PyJWT==2.5.0
python-decouple==3.6
SQLAlchemy==1.4.41
starlette==0.20.4
uvicorn==0.18.3
psycopg2==2.9.3